requirements.txt
# Web Framework
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-multipart==0.0.6

# Database
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.4

# Trading API
python-binance==1.0.17
pandas==2.0.1
numpy==1.24.3
ta==0.10.2  # Technical Analysis library

# Telegram Integration
python-telegram-bot==13.15

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Utilities
requests==2.29.0
python-dotenv==1.0.0
aiohttp==3.8.4
asyncio==3.4.3
websockets==11.0.3
schedule==1.2.0

# Testing
pytest==7.3.1
httpx==0.24.0